refactor: move interfaces as separate file
