+++ src/interfaces.ts
@@ -1,0 +1,8 @@
+interface IUser {
+  id?: string;
+  username: string;
+  age: number;
+  hobbies: Array<string>;
+}
+
+export { IUser};
